fix(scripts): use jq binary from node_modules in resolve:crcomlib (#25)
The resolve:crcomlib script will fail if the jq binary is not installed globally on the local machine. As node-jq installs a local copy of the jq binary into node_modules, this binary should be used so the script runs successfully regardless of whether the user has jq installed on their local machine or not.